requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-genai
requests
python-dotenv
gunicorn
